+++ FromGoldenCombs/BlockEntities/BELangstrothStack.cs
@@ -24,7 +24,7 @@ namespace FromGoldenCombs.BlockEntities
         int quantityNearbyHives;
         float _activityLevel;
         private RoomRegistry roomreg;
-        float roomness;
+        int roomness;
         public static SimpleParticleProperties Bees;
         int scanQuantityNearbyFlowers;
         int scanQuantityNearbyHives;
